fixed geocoding problems and availability bugs

+++ db/migrate/20200511041134_create_listings.rb
@@ -1,8 +1,8 @@
 class CreateListings < ActiveRecord::Migration[6.0]
   def change
     create_table :listings do |t|
-      t.bigint :longitude
-      t.bigint :latitude
+      t.float :longitude
+      t.float :latitude
 
       t.timestamps
     end
